app/auth.py
# app/auth.py 전체를 아래로 교체하세요
from datetime import datetime, timedelta
from typing import Optional
from jose import JWTError, jwt
import hashlib # passlib 대신 기본 hashlib 사용 시도
import os
from dotenv import load_dotenv

load_dotenv()

SECRET_KEY = os.getenv("SECRET_KEY")
ALGORITHM = os.getenv("ALGORITHM")

# 비밀번호 보안을 위해 간단하면서도 강력한 해싱 함수 정의
def get_password_hash(password: str):
    return hashlib.sha256(password.encode()).hexdigest()

def verify_password(plain_password, hashed_password):
    return get_password_hash(plain_password) == hashed_password

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    to_encode = data.copy()
    expire = datetime.utcnow() + (expires_delta or timedelta(minutes=1440))
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)